pyproject.toml
[project]
name = "VoteSystem"
readme = "README.md"
authors = [
    {name = "Arman Tovmasian", email = "armantovmasan14@gmail.com"},
]
requires-python = ">=3.10"
classifiers = [
    "Typing :: Typed",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["version"]

[project.optional-dependencies]
test = []
docs = [
    "mike >=1.1.0,<3.0.0",
    "mkdocs >=1.4.0,<2.0.0",
    "mkdocs-autorefs ==0.5.0",
    "mkdocs-click >=0.8.0",
    "mkdocs-include-markdown-plugin >=4.0.0,<7.0.0",
    "mkdocs-material >=9.0.0,<10.0.0",
    "mkdocs-material-extensions ==1.3.1",
    "mkdocstrings ==0.24.0",
    "mkdocstrings-python ==1.7.5",
]

[tool.poetry]
name = "VoteSystem"
version = "0.1.0"
description = "Voting system"
authors = ["Tovmasian Arman"]

[tool.poetry.dependencies]
python = "^3.10"
alembic = "1.13.1"
annotated-types = "0.6.0"
anyio = "4.3.0"
asyncpg = "0.29.0"
bcrypt = "4.1.2"
cffi = "1.16.0"
click = "8.1.7"
cryptography = "42.0.5"
dnspython = "2.6.1"
email_validator = "2.1.1"
fastapi = "0.110.0"
fastapi-users = "12.1.3"
fastapi-users-db-sqlalchemy = "6.0.1"
greenlet = "3.0.3"
h11 = "0.14.0"
httptools = "0.6.1"
idna = "3.6"
makefun = "1.15.2"
Mako = "1.3.2"
MarkupSafe = "2.1.5"
passlib = "1.7.4"
pycparser = "2.21"
pydantic = "2.6.3"
pydantic_core = "2.16.3"
PyJWT = "2.8.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.7"
pytz = "2024.1"
PyYAML = "6.0.1"
setuptools = "69.1.1"
sniffio = "1.3.1"
SQLAlchemy = "2.0.28"
SQLAlchemy-Ext = "0.2"
starlette = "0.36.3"
typing_extensions = "4.10.0"
gunicorn = "*"
uvicorn = "0.27.1"
uvloop = "0.19.0"
watchfiles = "0.21.0"
websockets = "12.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"



[tool.mypy]
no_incremental = true
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
allow_untyped_decorators = true
exclude = [
    "tests/",
    ".tmp/",
    ".venv/"
]